U.S sales dataset
Worked on the U.S. sales region dataset, performed data mining and modelling through power query and made new columns with the use of DAX.
Used bar, pie chart, slicer and did many visual formatting to make the layout more attractive and the content understandable.
Performed Data cleaning and mining in power query and removed blanked rows and null values. Replaced the row's name with a similar one to keep it simple. Made model relationships even more efficient for data flow. Did data visualization with the help of DAX and showed the total unit price and top and bottom customers, and used a pie chart to show the average percentage of unit price.
* Performed ETL ,Data cleaning and mining in power query.
* Made model relationship and used DAX functions for efficiency in power pivot.
* Used slicer to show year .
* Made bidirectional relationship.
* Shown top and bottom customers by topN filter and used bar charts in power view.
* Used geo map to show the sales in cities.
* Established pie chart to know the market cap.